+++ NEWS.md
@@ -16,6 +16,7 @@
 * Class of biodiversity data is identical for PO and PA
 * Bug fix in `built_formula_glmnet()` and response
 * Bug fix in `built_formula_gdb()` and response
+* Each model$biodiversity stores only predictors of current ID
 
 # ibis.iSDM 0.0.9